token manage to sdk
