+++ ecnu-openapi-sdk-csharp.csproj
@@ -8,11 +8,13 @@
     <Authors>ECNU</Authors>
     <Company>ECNU</Company>
     <PackageId>Ecnu.OpenApi.Sdk</PackageId>
-    <PackageReadmeFile>readme.md</PackageReadmeFile>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
     <Description>提供接入ECNU数据开发帮助，开发者仅需简单的配置即可获得ECNU授权的数据。</Description>
     <RepositoryUrl>https://github.com/ECNU/ecnu-openapi-sdk-csharp</RepositoryUrl>
-    <AssemblyVersion>1.0.1</AssemblyVersion>
-    <FileVersion>1.0.1</FileVersion>
+    <AssemblyVersion>1.0.2</AssemblyVersion>
+    <FileVersion>1.0.2</FileVersion>
+	<Version>1.0.2</Version>
+    <Title>提供接入ECNU数据开发帮助，开发者仅需简单的配置即可获得ECNU授权的数据。</Title>
   </PropertyGroup>
